change field type of ToleranceValue from i32 -> Decimal

+++ src/tolerance_value.rs
@@ -1,5 +1,7 @@
+use rust_decimal::Decimal;
+
 #[derive(Debug, Clone, Eq, Hash, PartialEq)]
 pub struct ToleranceValue {
-    pub(crate) upper: i32,
-    pub(crate) lower: i32,
+    pub(crate) upper: Decimal,
+    pub(crate) lower: Decimal,
 }
